#36 replace and rename tests
