+++ src/components/molecules/__tests__/FooterItem.test.ts
@@ -1,7 +1,7 @@
 ﻿import { describe, it, expect } from 'vitest';
 import { mount } from '@vue/test-utils';
 
-import FooterItem from '@/components/FooterItem.vue';
+import FooterItem from '@/components/molecules/MFooter.vue';
 
 describe('FooterItem.vue', () => {
   it('renders slots when passed', () => {
